+++ krypton/src/jsMain/kotlin/io/karma/evince/krypton/key/KeyPairGenerator.js.kt
@@ -17,16 +17,25 @@
 package io.karma.evince.krypton.key
 
 import io.karma.evince.krypton.Algorithm
+import io.karma.evince.krypton.Padding
 import io.karma.evince.krypton.annotations.UncheckedKryptonAPI
+import js.typedarrays.Uint8Array
+import web.crypto.CryptoKeyPair
+import web.crypto.KeyAlgorithm
+import web.crypto.KeyUsage
+import web.crypto.RsaHashedKeyGenParams
+import web.crypto.crypto
 
 /**
  * @author Cedric Hammes
  * @since  28/09/2024
- * @suppress
+ *
+ * @see [RSA-PSS, Web Cryptography API Examples](https://github.com/diafygi/webcrypto-examples?tab=readme-ov-file#rsa-pss---generatekey)
+ * @see [Web Cryptography API, W3C](https://w3c.github.io/webcrypto/#crypto-interface)
  */
 actual class KeyPairGenerator @UncheckedKryptonAPI actual constructor(
-    algorithm: String,
-    parameters: KeyPairGeneratorParameters
+    private val algorithm: String,
+    private val parameters: KeyPairGeneratorParameters
 ) {
     actual constructor(algorithm: Algorithm, parameters: KeyPairGeneratorParameters) :
             this(algorithm.validOrError(Algorithm.Scope.KEYPAIR_GENERATOR).toString(), parameters)
@@ -36,10 +45,27 @@ actual class KeyPairGenerator @UncheckedKryptonAPI actual constructor(
      * in the backend and the backend-internal structure is wrapped into a key.
      *
      * @author Cedric Hammes
-     * @since  26/09/2024
+     * @since  28/09/2024
      */
-    actual fun generate(): KeyPair {
-        TODO("Not yet implemented")
+    actual suspend fun generate(): KeyPair = (crypto.subtle.generateKey(
+        algorithm = when (algorithm) {
+            "RSA" -> when (parameters.padding?: Algorithm.RSA.defaultPadding) {
+                Padding.OAEP_SHA256 -> RsaHashedKeyGenParams.invoke(
+                    name = "RSA-OAEP",
+                    modulusLength = parameters.size,
+                    publicExponent = Uint8Array(arrayOf(1, 0, 1)),
+                    hash = KeyAlgorithm.invoke("SHA-256")
+                )
+                else -> throw IllegalArgumentException("")
+            }
+            else -> throw IllegalArgumentException("Algorithm '$algorithm' not supported")
+        },
+        extractable = false,
+        keyUsages = arrayOf(KeyUsage.decrypt, KeyUsage.encrypt)
+    ).unsafeCast<CryptoKeyPair>()).let {
+        KeyPair(
+            Key(algorithm, KeyType.PUBLIC, it.publicKey),
+            Key(algorithm, KeyType.PRIVATE, it.publicKey)
+        )
     }
-
 }